+++ packages/event_source_client/lib/src/sse_client_http.dart
@@ -4,6 +4,7 @@ import 'dart:math' as math;
 import 'package:http/http.dart' as http;
 
 import '../launchdarkly_event_source_client.dart';
+import 'events.dart' show isMessageEvent;
 import 'state_idle.dart';
 import 'state_value_object.dart';
 
@@ -18,7 +19,8 @@ class HttpSseClient implements SSEClient {
   static const defaultReadTimeout = Duration(minutes: 5);
 
   /// This controller is for the events going to the subscribers of this client.
-  late final StreamController<MessageEvent> _messageEventsController;
+  late final StreamController<Event> _messageEventsController;
+  late final EventSourceLogger _logger;
 
   /// This controller is for controlling the internal state machine when subscribers
   /// subscribe / unsubscribe.
@@ -37,9 +39,20 @@ class HttpSseClient implements SSEClient {
       Duration connectTimeout,
       Duration readTimeout,
       String? body,
-      String httpMethod)
-      : this.internal(uri, eventTypes, headers, connectTimeout, readTimeout,
-            _NoOpSink(), () => http.Client(), math.Random(), body, httpMethod);
+      String httpMethod,
+      EventSourceLogger? logger)
+      : this.internal(
+            uri,
+            eventTypes,
+            headers,
+            connectTimeout,
+            readTimeout,
+            _NoOpSink(),
+            () => http.Client(),
+            math.Random(),
+            body,
+            httpMethod,
+            logger);
 
   /// An internal constructor for injecting necessary dependencies for testing.
   HttpSseClient.internal(
@@ -52,8 +65,10 @@ class HttpSseClient implements SSEClient {
       ClientFactory clientFactory,
       math.Random random,
       String? body,
-      String httpMethod) {
-    _messageEventsController = StreamController<MessageEvent>.broadcast(
+      String httpMethod,
+      EventSourceLogger? logger) {
+    _logger = logger ?? NoOpLogger();
+    _messageEventsController = StreamController<Event>.broadcast(
       // this is triggered when first listener subscribes
       onListen: () => _connectionDesiredStateController.add(true),
       // this is triggered when last listener unsubscribes
@@ -73,16 +88,18 @@ class HttpSseClient implements SSEClient {
         random,
         body,
         httpMethod,
-        _resetRequest.stream));
+        _resetRequest.stream,
+        logger ?? NoOpLogger()));
   }
 
   /// Subscribe to this [stream] to receive events and sometimes errors.  The first
   /// subscribe triggers the connection, so expect a network delay initially.
   @override
-  Stream<MessageEvent> get stream => _messageEventsController.stream;
+  Stream<Event> get stream => _messageEventsController.stream;
 
   @override
   Future close() async {
+    _logger.debug('Closing SSE client permanently.');
     _messageEventsController.close();
     _connectionDesiredStateController.close();
     _resetRequest.close();
@@ -90,6 +107,7 @@ class HttpSseClient implements SSEClient {
 
   @override
   void restart() {
+    _logger.debug('Restarting SSE client.');
     if (_resetRequest.hasListener) {
       _resetRequest.sink.add(null);
     }
@@ -112,6 +130,7 @@ SSEClient getSSEClient(
         Duration connectTimeout,
         Duration readTimeout,
         String? body,
-        String method) =>
-    HttpSseClient(
-        uri, eventTypes, headers, connectTimeout, readTimeout, body, method);
+        String method,
+        EventSourceLogger? logger) =>
+    HttpSseClient(uri, eventTypes, headers, connectTimeout, readTimeout, body,
+        method, logger);