+++ packages/event_source_client/test/test_utils.dart
@@ -7,7 +7,8 @@ import 'dart:math' as math;
 import 'package:http/http.dart';
 import 'package:http/testing.dart';
 import 'package:launchdarkly_event_source_client/src/http_consts.dart';
-import 'package:launchdarkly_event_source_client/src/message_event.dart';
+import 'package:launchdarkly_event_source_client/src/logging.dart';
+import 'package:launchdarkly_event_source_client/src/events.dart';
 import 'package:launchdarkly_event_source_client/src/state_value_object.dart';
 
 class TestUtils {
@@ -24,11 +25,12 @@ class TestUtils {
       Duration? connectTimeout,
       Duration? readTimeout,
       Stream<bool>? connectionDesired,
-      EventSink<MessageEvent>? eventSink,
+      EventSink<Event>? eventSink,
       Sink<dynamic>? transitionSink,
       ClientFactory? clientFactory,
       math.Random? random,
-      Stream<void>? resetStream}) {
+      Stream<void>? resetStream,
+      EventSourceLogger? logger}) {
     return StateValues(
         uri ?? Uri.parse(defaultUri),
         eventTypes ?? defaultEventTypes,
@@ -36,13 +38,14 @@ class TestUtils {
         connectTimeout ?? Duration.zero,
         readTimeout ?? Duration.zero,
         connectionDesired ?? StreamController<bool>.broadcast().stream,
-        eventSink ?? StreamController<MessageEvent>.broadcast(),
+        eventSink ?? StreamController<Event>.broadcast(),
         transitionSink ?? StreamController<dynamic>.broadcast(),
         clientFactory ?? makeMockHttpClient,
         math.Random(),
         null,
         'GET',
-        resetStream ?? StreamController<void>.broadcast().stream);
+        resetStream ?? StreamController<void>.broadcast().stream,
+        logger ?? NoOpLogger());
   }
 
   static MockClient makeMockHttpClient(